Use physical size for resize buttons
Physical size and CSS size differ on high and low DPI devices.
Unless told otherwise JSO usees the physical size for the rendering
canvas, so when resizing the player element first convert the requested
size to the CSS size such that the resulting physical size and rendering
canvas match the requested dimensions.