refactor: rename rendermode identifiers
This better matches what the modes actually do.
Currently mostly internal apart from the lossy mode being referred to as
"fast" in the README, but the next commit will expose the internal names
to the public.
This does not change the existing option names, which are
already closer to the new than the old names.