Set default targetFps to 24
PAL and SÉCAM's exact 25fps and NTSC's (24000/1001)fps,
which is almost 24fps, are very common framerates for
eg anime releases (which anecdotally seem the most likely
to have complex typesets).
Some rendering modes deal (in case of lossy by design) very poorly with
falling behind the targeted rendering time deadline; lowering the
default from 30fps to 24fps increases the availalbe time per frame
by 25% and is closer to the common PAL, SÉCAM and NTSC framerates.

If data on the actual video framerate is available it is best to use
this to explicitly set the target framerate rather than relying on the
default.