Add option to limit canvas size
JSO struggles with complex subtitles on HiDPI canvasses,
so limiting the size can make sense if such subtitles are expected.

This was ported from jellyfin's fork and jellyfin-web uses this on at
least some devices. In the original implementation prescaleHeightLimit
would override maxRenderHeight if it was larger. This undocumentend
behaviour was dropped here as it seemed unintuitive and unhelpful.

[Vasily <just.one.man@yandex.ru>]
  Original implementation together with prescaling in
  https://github.com/jellyfin/JavascriptSubtitlesOctopus/commit/345d7010b1d31f933f1b2cc3a17d83a09b5e5c2e
  https://github.com/jellyfin/JavascriptSubtitlesOctopus/commit/4c5f018a07d5c51b089ef21ccf1f81a0355e6c8c
[Dmitry Lyzo]
  Rebased for upstream; seperated from the prescaling;
  removed default height limit and added docs
[Oneric]
  Renamed hardHeightLimit to maxRenderHeight;
  split from prescaling.

Co-Authored-By: Dmitry Lyzo <ashephard0@gmail.com>