Reject lossy renderMode if lacking createImageBitmap
Cherry-picked from
https://github.com/WeebDataHoarder/JavascriptSubtitlesOctopus/commit/2e79d26fe2fecedd305dcb3c570ad02f7d6d0ef2

Some older engines do not support the (Promise-based) createImageBitmap
API required for the lossy rendering mode. If that's the case force a
fallback to the default mode.
The original cherry-picked commit was adjusted for the renderMode
renaming and the second check outside the worker was dropped as we
use the relevant API exclusively inside the worker.