c++: add functions to detect and remove animatons
This prepares the dropAnimations and renderAhead features from
jellyfin, though the functions are not yet used. The implementation in
this commit completly replaces the faulty code from jellyfin, which was
added in the misleadingly named "Less zealous logging" commit
https://github.com/jellyfin/JavascriptSubtitlesOctopus/commit/6980debc4712239f2a3c481a3a6756eca2642aca

The old version was not detecting valid ASS tags like '\t  (\frz50'
and did some other checks not backed by ASS' behaviour. Rewrite it from
scratch to not miss any animated tags.
The previous version made an attempt to also detect and not report noop
animation tags like '\move(1,1,1,1)'; the new version does not as
correctly parsing the tag and its values is simply more effort than its
worth and likely to result in false negatives. Noop animation tags
should be sufficiently rare anyway.