make: consolidate link flags
All builds but brotli already set NO_EXIT_RUNTIME, and in brotli's case
it was probably an oversight for the flag to be missing.
All dependencies, but not the final jso linkage also set MODULARIZE,
but not -s 'EXPORT_NAME="someting"' as emscripten docs suggest should be
done. But since we only use static libs, the final ones are also
actually the only linking occuring, so just remove the flag.