src/make: use -c to compile instead of -r
The build currently warns about building an executable with an
object-file extension and recommends to use -c, -shared or -r
eventhough we already use -r.
I'm not sure what -r is supposed to do, but -c only compiles and not
links the input which should be exactly what we want here and using -c
gets rid of the warning without any apparent negative effects.