ci/gha: cache built dependencies
This speeds up builds significantly.
If there are any changes affecting cached files
but not updating the hash used as a key, increase
the prefix before the hash.

The uncoditional checkout of submodules even if they'll be replaced by
the cached version is necessary to generate .git/modules/lib/*/HEAD
used as a dependency for license generation.
We cannot just use the non-cached checkouts, since autoconf VPATH builds
do need to generate the configure script inside lib/*. The fresh
checkout doesn't have it, which will tirgger a rebuild.