Update emscripten to 3.1.15
EVAL_CTORS now errors when used without WASM,
previosuly it was likely just silently disabled.
Thus, only set it when linking with WASM.

There's now a new warning when closure isn't set explicitly
for non-WASM builds. Currently, there are errors concerning
redeclaration of the variable screen with closure enabled,
so explicitly disable it.
For the future, we probably want to remove the redeclaration
and enable closure.