Deprecate manual brotli decompression
It doesn't have much benefit but comes with
a maintenance cost as we either need to keep
patching upstream's js implementation for use in
JSO — which also bloats binary size by including
two copies of brotli and its default dict — or
use the C implemntation from JS which requires boilerplate
in both C and JS and also additional JS polyfills to keep
e.g. the fontname extraction working.

The browser's in-built support for Content-Encoding
is more felxible, faster and simpler to use.